requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
pydantic[email]
bcrypt
pyjwt
scikit-learn
# Optional: for LLM enhancement
langchain
langchain-groq